feat:지역 목록 조회 중간 저장

+++ S14P21A205_BE/src/main/java/com/ssafy/S14P21A205/store/repository/StoreRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
 
-    Optional<Store> findByUser_Id(Long userId);
+    Optional<Store> findByUser_Id(Integer userId);
 
     @Query("""
             select item.discountRate
